requirements.txt
torch==2.2.2
torchvision==0.17.2
Pillow==10.3.0
PyYAML==6.0.1
numpy==1.26.4
nibabel==5.2.1
timm>=0.9.16
